requirements.txt
flask
flask_restful
azure-identity
azure-servicebus